+++ src/test/java/com/bitbank/controllers/v2/AuthControllerTest.java
@@ -11,15 +11,19 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 
+import com.bitbank.dto.MfaDTO;
 import com.bitbank.entities.UsersEntity;
 import com.bitbank.repositories.UsersRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @AutoConfigureMockMvc
 @SpringBootTest
+@DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
 @TestPropertySource("classpath:application.properties")
 class AuthControllerTest {
 
@@ -46,6 +50,23 @@ class AuthControllerTest {
 		mvc.perform(post("/api/v2/auth/login").contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsBytes(userToLogin))).andExpect(status().isOk())
 				.andExpect(jsonPath("$.id", is("1")));
+	}
+
+	/**
+	 * Test invalid MFA exception.
+	 * 
+	 */
+	@Test
+	void testCanThrowInvalidMfaException() throws Exception {
+		BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+		String encodedPassword = passwordEncoder.encode("password");
+		// Insert an user
+		var user = UsersEntity.builder().username("username").password(encodedPassword).secretMfa("ABCDE").build();
+		usersRepository.save(user);
+		// Create the MFA
+		MfaDTO mfaDto = new MfaDTO("1", "123456");
+		mvc.perform(post("/api/v2/auth/verify-mfa").contentType(MediaType.APPLICATION_JSON)
+				.content(objectMapper.writeValueAsBytes(mfaDto))).andExpect(status().isUnauthorized());
 	}
 
 }